Update parse_pdb3.py
The distances are not calculated correctly. What we want to slide over the structure with a 3 residue window and compute the distance between C-alpha of the first and the third residues in that window, then slide again. Also, we don't want all phi and psi angles. Compute them only for windows that the calculated distance is below a threshold (let say 3.5 A).